requirements.txt
lakefs_client==0.69.1
pydantic
fastapi
pyyaml
uvicorn[standard]
httpx
boto3